CP-24350: use structured logging instead of encoding data in message
This replaces almost every instance of the Msgf() calls with Msg();
we now use structured logging instead of using a format string to
encode values into the message.

This makes the logs easier to work with; every instance of the log
has the same message, and we can work with the logs without unpacking
the relevant data from the message, which makes more advanced queries
possible.

I've also tweaked a few instances where we weren't using Msgf, but
had `log.Error().Err(err)....`, which can be simplified to
`log.Err(err)....`.