Fix file_monitor test on Linux
On macOS, all we get when creating a new file with contents is a
FileCreated event, but on Linux there is a FileCreated event and a
FileChanged event (the file is first created, then the contents
written). This can be replicated by running the tests in a Linux
container.

This just adds the FileChanged event to the test so the mock knows
it's not an unexpected event.