CP-25629: add Helm chart for deploying collector/shipper, fix up docs.
As I was trying to get the collector/shipper installed on my cluster, I
ran into some difficulties with the instructions; a lot of stuff was
hard-coded, and there were a few inconsistencies. I figured the easiest
thing to do would just be to throw the Kubernetes stuff into a Helm
chart, so I did.

For the non-Kubernetes portions (e.g., `eksctl` commands), I've tried
to clean everything up and parameterize it so it can be used with
multiple clusters.