CP-25372: switch to Brotli-compressed JSON for on-disk storage
Unsurprisingly, this has two parts: the collector and the shipper.
The collector side is pretty straightforward; instead of writing
Parquet files, we write Brotli-compressed JSON.

Things are a bit more interesting on the shipper side, as we still
want to upload Parquet files to S3. Therefore, I've tweaked the
`shipper.File` struct to automatically transcode the *.json.br files
to Parquet as they're read.

In order to save memory, I implemented this as a streaming operation.
However, `File` currently caches a file in memory, and I didn't want
to be /too/ invasive with this patch, so we still do that with this
patch. However, in the future I believe this represents an good
opportunity to reduce our memory usage, as we could just pass the
stream directly to the S3 uploader and avoid having to keep a copy
of the entire Parquet file in-memory.