Split in-memory Metric type from storage formats
This allows us a lot more flexibility with different formats since we
no longer need to use the same Go struct for multiple formats and can
instead use custom marshal/unmarshal functions to get exactly what we
want. Our in-memory and JSON types are now more compact and idiomatic,
while the Parquet remains unchanged.

This also centralizes some of the logic around the label <-> field
conversion in the Metric type so we can share it across a few different
places without reimplementing it, which is particularly good if we're
planning on expanding that to other fields/labels in the future.