CP-27103: fix gosec G115 warnings.
G115 was warning about overflows in CI (Linux) but not on our laptops
(macOS). The problem was that the types of the numbers we are operating
on varied; syscall types are highly platform-specific.

On Linux, some of these were int64 whereas on macOS they are actually
uint64. This meant that the gosec G115 warning would be triggered on
Linux, where we were mixing uint64 and int64 types. Adding a //nolint
directive silenced the error on Linux, but since it wasn't required on
macOS the nolintlint linter was triggering.

The fix just uses the safecast module to cast whatever the underlying
type is to a uint64. It will panic if it fails (since I used the
`MustConvert` function, not `Convert`), but I think that's the right
thing to do here... I'd rather see a panic here than a non-trapping
overflow.