CP-27359: use `go list -m` to get module name in build system
This is a *lot* more reliable than the previous method, since it will
give us the Go module name (based on go.mod) instead of the directory
name from which we'll need to piece together the full repo name.  It
also allows us to drop the dependency on git when building the Docker
image.

I did a bit more cleaning up in the Makefile than was strictly
necessary, but since I was touching a bunch of this stuff anyways I
figured I'd take the opportunity to clean up a few variable names.

I also removed some duplicate ldflags; we were previously setting each
of the build variables twice, once based on the repo name (based on the
directory name) and again hardcoded to cloudzero-insights-controller.
Since we now have the proper Go module name, we can ditch the repo name
stuff for something we know will work, and drop the duplicate hardcoded
properties.