CP-27063: make entire destination URL configurable
Previously, we only had a `host` property which was configurable
Technically you could actually provide a path as well, but the code
would always append /v1/container-metrics to the path.  Since we
want to direct traffic at the collector we need the path to be
/collector.

This allows us to configure not only the host and path, but also the
scheme, while preserving the default of sending data to the CloudZero
/v1/container-metrics endpoint.

Note that this is techincally a backwards-incompatible change. However,
we can preserve compatibility in the Helm chart, which is really the
only place this should currently be invoked.