CP-27611: unify code for referencing images
This adds an agent.images section, which contains default values for
all images used in the Helm chart, as well as a helper function which
accepts an image struct (which we use in many places) and falls back
on the defaults from agent.images if a property isn't set.

Once we have made a release from the agent-validator repo, we should
be able to delete *most* of the values from the image configuration
sections and have them all use the default values automatically, though
overrides set by customers will still be honored.