CP-27533: change repository references to cloudzero-agent
Once we have renamed the repository from cloudzero-agent-validator
to cloudzero-agent, we will need to update many references to use
the new repository name.

Unfortunately we can't just naively replace cloudzero-agent-validator
with cloudzero-agent as cloudzero-agent-validator is also the name
of an executable in this repo. However, replacing the full URL
(github.com/cloudzero/cloudzero-agent-validator) in the Go files does
*most* of the work for us. The rest is just a matter of going through
occurances of "cloudzero-agent-validator" and switching the ones which
refer to the repository but not the ones that refer to the executable.