CP-27297: move metrics settings out of values.yaml
We don't want to expose the list of metrics we collect in the Helm
chart for users to configure. This moves them into a template file
instead.

The syntax is a bit uglier since we have to include the template
than pipe it to fromYaml, but it does accomplish the goal.