CP-27353: remove pkg/build
This removes the final piece of pkg/, but unlike all the other changes
where I just moved the packages around without modification, for this
one I decided to update the code to use the same app/build other apps
use.

The major change is that I removed the application name from the build
configuration, and the GetVersion function. Previously, it was hard-
coded to cloudzero-insights-controller, which was wrong for *most* of
the applications anyways.