CP-28160: add Helm template generation test
This adds a target to the Makefile to automatically regenerate a
manifest as part of the `generate` Make target. The output goes to
tests/helm/template/manifest.yaml, which is checked into the repo.

The purpose of this change is to make sure that every change to the
output of the Helm chart is intentional; if you alter the Helm chart
without a corresponding alteration of the generated output, a status
check should fail. Due to the fact that the manifest is included in
the repo, the effect of the changes should be apparent in the commit,
and will be easy to notice during a PR review (if not sooner).