CP-28572: regenerate Helm template tests when making a release.
When making a release, we automatically bump the version of the image
we refer to in the Helm chart to release of cloudzero-agent that we
are creating. This obviously causes a change in the generated manifest,
which we didn't previously account for. This just regenerates the
templates in order to keep them in sync with the Helm chart.