CP-28746: fix domain for webhook service
In some situations, the webhook-configuration configmap would contain
an incorrect value for the domain, causing the backfill job to fail
due to an inability to resolve the (incorrect) DNS entry for the
webhook service.

This patch just tweaks the helper to generate the name using the same
function which is used to generate the service name in the service
resource.