CP-27300: clean up file organization of Helm chart
Previously, we were a bit haphazard about where we put stuff; mixing
aggregator and webhook stuff in the same file, etc.

This just reorganizes the files a bit. Each resource is now in its own
file, prefixed with the part of the chart (webhook, agent, aggregator,
etc.) with a suffix indicating the resource type.

This shouldn't change anything functionally, it's just a reorganization
of existing resources. Unfortunately the manifest.yaml is a bit
difficult to grok because stuff moved around.