CP-29142: complete Helm schema validation.
This adds a lot of additional validation to the values.yaml schema,
as well as some supporting infrastructure. Specifically, lots of tests
and the schema is now written in YAML, which is automatically processed
to JSON (as Helm expects) in the Makefile.

This provides a couple of big advantages. First, it documents the
format of a lot of stuff in the Helm chart, which was honestly a bit
tricky to figure out.

Because Helm will perform validation against the schema by default, it
also allows us to catch many configuration errors much sooner. In
particular, because we use `additionalProperties: false` pretty much
everywhere, it protects against users providing configuration which
isn't actually used anywhere, but that they expect to be.

This also provides a significantly improved upgrade path for customers
who are currently relying on deprecated, unstable APIs... at some
point in the future, we can simply remove the deprecated items from
the schema and customers will get an error message instead of their
configuration overrides just silently no longer working.

This also add several missing properties to values.yaml and schema.
These properties are referenced in the Helm chart, but were not
previously mentioned in the values.yaml, nor included in the schema.