WIP: Significantly expand Helm schema validation
This adds a lot of additional validation to the values.yaml schema,
as well as some supporting infrastructure. Specifically, lots of tests
and the schema is now written in YAML, which is automatically processed
to JSON (as Helm expects) in the Makefile.

This isn't done yet; I still need to test a bunch of things, and more
carefully examine the allowed types in the aggregator top-level config
(components.aggregator should be good).