Add maybeGenerateSection template
This is just a helper that will allow you to specify some data and,
if the data is not empty, output a section in the YAML. This allows
us to change things like

```yaml
{{ if $foo -}}
foo:
  {{ $foo }}
{{- end }}
```

to

```yaml
{{- include "cloudzero-agent.maybeGenerateSection" $foo -}}
```

I've also switched over a few of the existing templates to make use of
this.

I'm honestly not sure where exactly this came from; I think maybe from
when I was playing with Checkov. It was just sitting around in my repo
uncommitted, but it's a good change (IMHO) so I figured I'd just throw
up a quick PR.