CP-29051: attempt to fix flaky lock file test
I can't reproduce this locally, but sometimes (rarely) in CI we see
a failure in this test:

TempDir RemoveAll cleanup: unlinkat /tmp/TestLock_MaxRetry.../001: directory not empty

I'm *hoping* this resolves the issue. I'll also be very curious to
to see if we see any errors from the Release call that wasn't
previously checked.