CP-29129: remove obsolete section on ArgoCD (#238)
These annotations were necessary in days of yore, when we had a
`ttlSecondsAfterFinished` on the jobs. However, we removed that a while
back in favor of just giving each job a unique name derived from a hash
of the configuration version. With that change, the jobs are run
whenever the configuration is updated without relying on the cluster to
delete the old jobs.