CP-27814: remove unnecessary configurability
Since this is a new component, we don't need to retain backwards
compatibility with all of the server values. This allows us to
remove some complexity from the chart.

There are also a few cleanups to remove vestigial properties which
are no longer used, and a few minor cleanups for comprehensibility.