CP-29051: if Go tests fail in the merge queue, rerun them (#242)
A few of our Go tests are a bit flaky. Normally this isn't a major
problem since it's pretty straightforward to re-run the failed tests,
and honestly it's a good reminder to fix those flaky tests.

However, with the merge queue this is extremely annoying, since if
*any* test fails it gets kicked out of the queue and we have to re-add
it manually, at which point it runs *all* the tests again.

So I thought it might be a good idea to tweak the CI setup a bit so
that, if we're running in the merge queue, `make test` failures will
simply be re-run (up to thrice). However, if we aren't in the merge
queue only make a single attempt.

Hopefully this becomes less important as time goes on and we are able
to fix these tests, but in the meantime this should be quite useful.

I'm not certain the branch name is correct, but if not I'll at least
have the correct name in the future thanks to the dump-context step.