Remove unnecessary properties from Prometheus configuration.
These all just set the value to the same value as the default, which
just clutters up the configuration.

The only exception is the scrape_protocols, where the default also
contains `PrometheusProto`, which is the newest (and supposedly most
efficient) protocol. Based on internal conversations, it sounds like
the initial code for these configurations was copied from an example
which was likely written before PrometheusProto existed.