CP-29585: periodically close HTTP connections
When Prometheus sends data to the Collector, it uses the standard Go
HTTP client library, which reuses the connection as an optimization.

Unfortunately, the round robin to distribute the load across the k8s
service happens when the initial connection is made, and since in
non-federated mode we really only receive data from Prometheus and the
Webhook server, which means we just have two things that create an
initial connection, then always send data over those connections.

Luckily, and somewhat surprisingly (at least to me), Prometheus seems
to be using HTTP/1.1 to send the data, which means we can just add a
Connection: close periodically to get it to close the connection and
go through the load balancing again.