CP-29706: run `format` make target after regenerating manifests (#286)
I'm not entirely sure why, but `make` detects that
helm/values.schema.json needs to be regenerated when running in CI.
This is handled automatically in the Makefile, but reformatting the
result (using prettier) is not.

Since we should already have prettier installed due to the
`install-tools` target, we can just add the `format` target when
running `helm-generate-tests` and reformat the output before
committing. This should basically turn the regeneration of the
JSON schema into a no-op, but without the release-to-main workflow
will fail due to uncommitted changes in `helm/values.schema.json`.