CP-29735: allow 'global' top-level property, for subcharts (#288)
When the cloudzero-agent Helm chart is used as a subchart, Helm will
automatically add a top-level 'global' property (for detais, see
https://helm.sh/docs/chart_template_guide/subcharts_and_globals/).

What I didn't anticipate is that Helm will also include that property
in the JSON Schema it is checking against, meaning that previously any
use of the cloudzero-agent chart as a subchart would cause a schema
validation error.

This tweaks the JSON Schema to allow for a 'global' property, with any
contents. It also adds a test which would have caught this issues, to
ensure that there are no regressions.