CP-29742: add a helmless job (#290)
This will just run helmless in a job when the chart is installed,
and we can check the logs in order to get the minimal overrides for
the installation.

This also moves the config-values ConfigMap to helmless-cm, and alters
Helmless to embed a copy of the defaults in the tool.