CP-29491: don't clobber defaults when merging component-specific config (#297)
The `merge` function in Helm will write to the destination parameter
(the first parameter), so if we just pass the defaults directly and
merge into them, we will end up adding the values for the component
we're merging to the defaults.

Instead, we use the `deepCopy` function to create a copy of the
defaults, then merge into that.