CP-29986: add additional properties to schema validation & values.yaml (#311)
* CP-29986: add additional properties to schema validation & values.yaml

These properties are used in the Helm chart, but were not present in
values.yaml or the JSON Schema (which was based on values.yaml).

This patch adds the properties to values.yaml, the JSON Schema, and
adds a bunch of tests to validate many of the new properties.

* Remove superfluous properties from schema tests.

These are provided by the test fixture, so including them in the tests
is not necessary.