Refactor helm-test-schema make target.
Previously, we had one target which would iterate through all of the
tests, running both `helm template` and `kubeconform`. This worked
well, but was slow.

This change creates multiple targets for each test, one for kubeconform
and another for helm template, then changes the helm-test-schema target
to simply depend on those targets. This allows us to execute the tests
in parallel, with something like `make -j helm-test-schema`.