CP-29693: add "scout" package for detecting CSP metadata
The Scout is indended to gather information about the environment.
Currently, this means determining which Cloud Service Provider's
infrastructure the code is running on, and determining the region and
account ID.

This patch contains several components:

 * `cloudzero-scout` executable,  which can be useful for debugging,
   but is not used during normal operations.
 * `scout` Go package is where the actual implementation resides.
 * Changes to configuration packages to use the scout where
   appropriate (i.e., if a cloudAccountId or region are not provided).
 * JSON Schema changes to allow not passing a cloudAccountId or region.

If the cloudAccountId or region are not provided, and are not able to
be detected, a fatal error will occur.