CP-29985: fix Webhook certificate annotation handling (#312)
Previously, we wrote the correct annotations in the wrong spot, and
the incorrect annotations in the right spot.

This just changes the template to output the right annotations at the
right location, and also adds a template test for cert-manager to
help us keep track of certificate-related changes in the future.