CP-30020: add kubeconform tests (#314)
* CP-30020: use kubeconform to validate our generated templates

Crucially, this is integrated into the schema validation (the
helm-test-schema make target), so we're exercising a lot of different
properties here that don't get used very often.

* Fix issues identified by new kubeconform tests.

* Refactor helm-test-schema make target.

Previously, we had one target which would iterate through all of the
tests, running both `helm template` and `kubeconform`. This worked
well, but was slow.

This change creates multiple targets for each test, one for kubeconform
and another for helm template, then changes the helm-test-schema target
to simply depend on those targets. This allows us to execute the tests
in parallel, with something like `make -j helm-test-schema`.

* Add GitHub Action job to run helm-test target.