CP-30242: import only the definitions we use from K8s JSON Schema
The previous implementation imported all 707 definitions from the
Kubernetes JSON schema file, even though the Helm values schema only
references a small subset. This resulted in unnecessarily large output
files.

This first updates the JSON Schema to the latest version.

Next, this modifies the merge-json-schema.jq script to only import
definitions that are actually referenced. This reduces the imported
definitions from 707 to 129 (81.8% reduction) and output file size from
1.2M to 311K.

It also removes the "description" fields, which Helm doesn't make use
of, further reducing the file size to 97K for a total of a 92%
reduction in file size.