CP-30258: rework validation of PDBs
`minAvailable` and `maxUnavailable` cannot both be set; Kubernetes will
error if they are.

Since we set `minAvailable` to 1 by default, if a user sets
`maxUnavailable` without explicitly setting `minAvailable` to null,
it will result in an invalid chart, which can be counter-intuitive.

This adds some extra validation to the generatePodDisruptionBudget
helper function which detects cases where both are set, and emits
an error explaining not just that both cannot be set, but that due
to the default value if you use `maxUnavailable` you need to explicitly
set `minAvailable` to null.

It also changes how we handle overriding PDBs an the component level;
instead of merging the specs at the propertly level like we do for
other properties, we look to see if either property (`minAvailable` or
`maxUnavailable`) is set and, if so, completely override the default
configuration. This removes the need for the user to explicitly set
`maxAvailable` to `null` when specifying `minUnavailable` in an
override.