CP-30274: allow re-releasing (#337)
Sometimes something goes wrong with a release, and we want to
re-release after fixing. Unfortunately, the workflow to bump the
version number in the chart will fail in this case.

This just updates the workflow to account for when setting the version
number to the new release is a no-op.