CP-30309: add *.{yaml,json}-semdiff targets
These perform a semantic diff against the specified files to see if
there are any differences beyond just formatting noise. This is
extremely useful for refactoring the Helm charts.

This also adds a helm-test-template-diff target, which is a shortcut
for running the *-semdiff target on all of the template tests.