CP-30309: add *.{yaml,json}-semdiff targets (#343)
* Use strict version of Kubernetes JSON Schema

The strict version is basically indentical, except that everything has
additionalProperties set to false. This matches our newish strategy
of trying to detect misconfigurations earlier in the process.

* Run files through prettier as they are generated.

Instead of waiting for a separate formatting pass (i.e.,, the
`format-prettier` target), this will just pipe data through prettier
as it is generated. Hopefully this will eliminate some friction in
the development process.

* Improve file-based targets and dependencies in Makefile

We have been relying on phony targets a bit too heavily. Switching
to a file-based approach in many places should be a big win:

* Allows more granularity in terms of exactly what we doing.
* Much easier to automatically rebuild generated dependencies.
* Significant speed-ups when using parallel builds.

This also moves mock genenation out of go:generate directives and
into the Makefile so they are easier to keep up to date.

* Introduce KUBE_VERSION variable in Makefile, use it.

This just cuts down on repetition a bit.

* Rename helm-generate-tests to helm-tests-template.

This is just for consistency with the other helm-test-* targets.

* Use local versions of tools, instead of relying on setting PATH

The PATH stuff seems to be somewhat unreliable; I'm not sure exactly
why, but it was using global mockgen instead of the local one.

* CP-30309: add *.{yaml,json}-semdiff targets

These perform a semantic diff against the specified files to see if
there are any differences beyond just formatting noise. This is
extremely useful for refactoring the Helm charts.

This also adds a helm-test-template-diff target, which is a shortcut
for running the *-semdiff target on all of the template tests.