CP-31738: switch from bitnami/kubectl to d3fk/kubectl
Broadcom/VMWare seems to be enganging in a bit of a money-grab,
and are discontinuing most of the Bitnami images in favor of
"Bitnami Secure Images"; see
https://news.broadcom.com/app-dev/broadcom-introduces-bitnami-secure-images-for-production-ready-containerized-applications

These new images are apparently more hardened, but non-free.
Some of the new images do have free versions available, but
only the "latest" tag, so they're not really suitable for
inclusion in this chart.

The d3fk/kubectl repository looks to be actively maintained,
with new images being automatically generated on a regular
basis.