CP-32687: detect missing tables during update/insert and treat as fatal (#452)
We have seen a couple of instances where the SQLite tables don't seem
to be created during initialization, leading to an infinite stream of
errors when we attempt to write data to the table.

This patch will simply detect these errors, and change the log event
from a warning to a fatal, causing the process to exit, which will
give Kubernetes a chance to restart the pod and initialize a new
database, this time hopefully *with* the missing table(s).