CP-33387: use protoc installed by arduino/setup-protoc action (#479)
We have been getting sporadic errors from CI for a while now due to
a version difference in protoc (it embeds the version number the files
were generated with in the source code, so when it changes CI detects
the difference and, correctly, fails).

This tweaks the workflow ever so slightly to make sure that the protoc
we invoke is the one setup-protoc installs. Hopefully this will clear
up those sporadic errors.